Add Topic tab and redesign project overview
